Ensure nauty build failures are detected properly
By using `&&` to chain the commands in the build script, a failure in
any script (e.g. in `configure`) causes the whole thing to abort, and
report a non-zero exit code.

Previously, if e.g. configure failed, the build processes nevertheless
proceeded, and ultimately always reported success. This is annoying for
continuous integration, but also can be confusing for users.